+++ src/core/dom/image/test/index.js
@@ -209,6 +209,23 @@ module.exports = async (page, params) => {
 					expect(await getNode(tag).evaluate((ctx) => globalThis.getSrc(ctx))).toBe(images.pngImage2x);
 				});
 
+				it('update `src` without ctx provided', async () => {
+					await imageLoader.evaluate((imageLoaderCtx, [tag, images]) => {
+						const target = document.getElementById(`${tag}-target`);
+						imageLoaderCtx.init(target, {src: images.pngImage, handleUpdate: true});
+					}, [tag, images]);
+
+					await h.bom.waitForIdleCallback(page);
+
+					await imageLoader.evaluate((imageLoaderCtx, [tag, images]) => {
+						const target = document.getElementById(`${tag}-target`);
+						imageLoaderCtx.update(target, {src: images.pngImage2x, handleUpdate: true});
+					}, [tag, images]);
+
+					await h.bom.waitForIdleCallback(page);
+					expect(await getNode(tag).evaluate((ctx) => globalThis.getSrc(ctx))).toBe(images.pngImage2x);
+				});
+
 				it('with `src` and preview with `src`', async () => {
 					const
 						imgUrl = getRandomImgUrl(),