+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.0.0-rc.141 (2021-02-05)
+
+#### :bug: Bug Fix
+
+* Fixed an issue with an update without providing of `ctx` `core/come/image`
+
 ## v3.0.0-rc.140 (2021-02-05)
 
 #### :rocket: New Feature